posix-aio-compat.c: remove compiler warnings.

+++ posix-aio-compat.c
@@ -205,7 +205,7 @@ static ssize_t handle_aiocb_rw_linear(struct qemu_paiocb *aiocb, char *buf)
     ssize_t offset = 0;
     ssize_t len;
 
-    while (offset < aiocb->aio_nbytes) {
+    while ((size_t)offset < aiocb->aio_nbytes) {
          if (aiocb->aio_type & QEMU_AIO_WRITE)
              len = pwrite(aiocb->aio_fildes,
                           (const char *)buf + offset,
@@ -252,7 +252,7 @@ static ssize_t handle_aiocb_rw(struct qemu_paiocb *aiocb)
          */
         if (preadv_present) {
             nbytes = handle_aiocb_rw_vector(aiocb);
-            if (nbytes == aiocb->aio_nbytes)
+            if (nbytes == (ssize_t)aiocb->aio_nbytes)
                 return nbytes;
             if (nbytes < 0 && nbytes != -ENOSYS)
                 return nbytes;
@@ -447,7 +447,7 @@ static int posix_aio_process_queue(void *opaque)
                 /* end of aio */
                 if (ret == 0) {
                     ret = qemu_paio_return(acb);
-                    if (ret == acb->aio_nbytes)
+                    if (ret == (int)acb->aio_nbytes)
                         ret = 0;
                     else
                         ret = -EINVAL;
